Use the author date, not committer date for the bup save date.
Since "bup get" is going to change the committer date for fetched
commits to match the current date, use the author date to represent
saves so that fetched saves will have the same names as the originals.

This should have no effect with respect to backward compatibility
because until now, bup always used the same date for the author and
committer.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>