Multiplex stdout and stderr over stderr for "bup on".
Previously, when running "bup on" (i.e. when in BUP_SERVER_REVERSE
mode), bup would redirect the remote command's stdout to stderr
because stderr is the only remaining avenue back to the local console,
given that in reverse mode, stdout (and stdin) are connected back to a
local "bup server" instance (hence the name).

Of course that makes it impossible to reliably capture the
non-diagnostic output from the remote commands, i.e.

  $ commit_id="$(bup on HOST save -t ...)"

To fix that, have all "bup on" subcommands encode their stdout and
stderr output via TaggedOutput instances whenever they detect that
they're running in reverse.

With the current trivial protocol, the end result is that all remote
commands will write lines to stderr that look like this:

  on This is a stdout line that had a newline terminator.
  o_ This is a stdout line that didn't.
  en This is a stderr line that had a newline terminator.
  o_ This is a stderr line that didn't.

The loop in on-cmd.py handles demultiplexing this output (from the
remote stderr) to the local stdout/stderr.

Questions:

Right now TaggedOutput does line-buffering, which makes things more
complicated -- is that worthwhile?

And the interaction with _hard_write() is messy -- I wonder if we
still need _hard_write() (i.e. has ssh still broken), and if so,
exactly what the expectations are, i.e. is *all* stderr output
expected to go through _hard_write(), or just log()?

Signed-off-by: Rob Browning <rlb@defaultvalue.org>