Added command "bup log"
"bup log" can be used to inspect the backup history.
It is a wrapper for "git log" and "git diff-tree" and uses them in a
way that is convenient for the bup use case. Furthermore, "bup log"
knows about name mangling and will hide the name mangling from the
user, a feature that "git log" cannot provide directly.
The man-page contains detailed information and examples.

Signed-off-by: Holger Dell <github@holgerdell.com>